+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/Proxmox/VdcHome.ascx.cs
@@ -180,7 +180,7 @@ namespace SolidCP.Portal.Proxmox
 
         protected void ddlPageSize_SelectedIndexChanged(object sender, EventArgs e)
         {
-            gvServers.PageSize = Convert.ToInt16(ddlPageSize.SelectedValue);
+            gvServers.PageSize = Convert.ToInt32(ddlPageSize.SelectedValue);
             gvServers.DataBind();
         }
