+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/SpaceMenu.ascx.cs
@@ -82,7 +82,7 @@ namespace SolidCP.Portal
                         Session["currentUser"] = PanelSecurity.SelectedUserId;
                     }
                 }
-                currentPackage = Convert.ToInt16(Session["currentPackage"]);
+                currentPackage = Convert.ToInt32(Session["currentPackage"]);
             }
             else
             {
