+++ SolidCP/Sources/SolidCP.WebPortal/DesktopModules/SolidCP/VPS/VdcHome.ascx.cs
@@ -143,7 +143,7 @@ namespace SolidCP.Portal.VPS
 
         protected void ddlPageSize_SelectedIndexChanged(object sender, EventArgs e)
         {
-            gvServers.PageSize = Convert.ToInt16(ddlPageSize.SelectedValue);
+            gvServers.PageSize = Convert.ToInt32(ddlPageSize.SelectedValue);
             gvServers.DataBind();
         }
